JS tests do not write to fixture file anymore and use helper methods form separate helper module
