+++ node_package/tests/vm.test.js
@@ -1,18 +1,15 @@
 const test = require('tape');
-const path = require('path');
 const fs = require('fs');
+const { getUploadedBundlePath, createUploadedBundle } = require('./helper');
 const { buildVM, runInVM, getBundleUpdateTimeUtc } = require('../src/worker/vm');
 const { initiConsoleHistory } = require('../src/worker/consoleHistory');
-
-function getBundlePath() {
-  return path.resolve(__dirname, './fixtures/bundle.js');
-}
 
 if (!console.history) initiConsoleHistory();
 
 test('buildVM and runInVM', (assert) => {
   assert.plan(2);
-  buildVM(getBundlePath());
+  createUploadedBundle();
+  buildVM(getUploadedBundlePath());
 
   assert.deepEqual(runInVM('ReactOnRails'), { dummy: 'Dummy Object' }, 'ReactOnRails object is availble is sandbox');
   assert.ok(global.ReactOnRails === undefined, 'ReactOnRails object did not leak to global context');
@@ -20,7 +17,8 @@ test('buildVM and runInVM', (assert) => {
 
 test('VM security', (assert) => {
   assert.plan(3);
-  buildVM(getBundlePath());
+  createUploadedBundle();
+  buildVM(getUploadedBundlePath());
 
   // Adopted form https://github.com/patriksimek/vm2/blob/master/test/tests.js:
   assert.throws(
@@ -41,7 +39,8 @@ test('VM security', (assert) => {
 
 test('VM console history', (assert) => {
   assert.plan(2);
-  buildVM(getBundlePath());
+  createUploadedBundle();
+  buildVM(getUploadedBundlePath());
 
   let vmResult = runInVM('console.log("Console message inside of VM") || console.history;');
   const consoleHistory = [{ level: 'log', arguments: ['[SERVER] Console message inside of VM'] }];
@@ -60,9 +59,11 @@ test('VM console history', (assert) => {
 
 test('getBundleUpdateTimeUtc', (assert) => {
   assert.plan(1);
-  buildVM(getBundlePath());
+  createUploadedBundle();
+  buildVM(getUploadedBundlePath());
+
   assert.equal(
     getBundleUpdateTimeUtc(),
-    +(fs.statSync(getBundlePath()).mtime),
+    +(fs.statSync(getUploadedBundlePath()).mtime),
     'getBundleUpdateTimeUtc() should return last modification time of bundle loaded to VM');
 });
